renamed file to CustomerReport
